refactor(app): extract SpeakerNamingSession from PipelineQueue
Move the speaker-naming half of the pipeline (the possibly-late naming
dialog, its persisted sidecars + recognition forensics, and the late
re-diarization / re-apply paths) out of the ~2050-line PipelineQueue into
a dedicated @MainActor @Observable SpeakerNamingSession collaborator, split
across SpeakerNamingSession.swift and SpeakerNamingSession+Late.swift.

Why a same-isolation delegate, not an actor: the naming work reads and
mutates queue-owned job state on the main actor with no suspension between
the reads and the mutation the RPC contract depends on. The queue owns the
session strongly (let naming) and sets naming.delegate = self post-init;
the session holds the delegate weakly, so there is no retain cycle and a
released queue turns every callback into a no-op via optional chaining.

The delegate protocol is strictly id-based, never index-based: a job index
goes stale across an await when another finished job is evicted by
completedJobLifetime (a documented past bug), so the session addresses jobs
only by UUID.

Behaviour is preserved byte-for-byte:
- The synchronous .speakerNamingPending -> .generatingProtocol hop inside
  completeSpeakerNaming stays synchronous (the /v1/jobs/<id>/naming
  idempotency contract).
- The updateSpeakerDB write + refreshKnownSpeakerNames pairing stays on the
  queue as one atomic delegate call (issue #155 CPU-regression guard).
- The SpeakerNamingData Codable shape and its revision UUID change-detection
  marker are moved verbatim; the types stay nested under PipelineQueue (via
  an extension) so every PipelineQueue.SpeakerNamingData reference across the
  UI, RPC, persistence, and tests keeps resolving.

The queue keeps thin get/set forwarders (speakerNamingDataByJob,
speakerNamingHandler, completeSpeakerNaming, cleanupStalePending,
pendingSpeakerNaming, speakerNamingData(forJobID:)) so all external call
sites and SwiftUI observation through queue.naming keep working unchanged.
DiarizationRun becomes an internal top-level type since it crosses the
queue <-> session boundary in both directions.

PipelineQueue.swift: 2043 -> 1534 lines. New session files stay under the
600-line file_length cap; the queue keeps its existing suppressions (they
drop in a later step).
